ml/requirements.txt
requests
google-cloud-bigquery
google-cloud-language
google-cloud-storage
google-cloud-speech
ipython==5
